Updated to Margin Info Tile Card
Minor UI update to provide more actionable information on Margin Collateral Information.

Collateral references have been updated to display in $ terms and precise margin % (to 10,000% for handling of dust borrows) as this is more actionable and has been requested by quite a few users. There is still a reference to the minimum MCR % in the tooltip.

This would be an interim step to a full featured liquidation calculator, but should be helpful to users whilst that is in development.

Please let me know if any updates are required.